+++ src/main/java/com/itesm/panoptimize/dto/agent/AgentCreateDTO.java
@@ -10,7 +10,7 @@ public class AgentCreateDTO {
     private String email;
     @NotNull(message = "FullName is required")
     private String fullName;
-    private String imagePath;
+
     @NotNull(message = "RoutingProfileId is required")
     private String routingProfileId;
     private boolean canSwitch = true;
@@ -46,13 +46,7 @@ public class AgentCreateDTO {
         this.fullName = fullName;
     }
 
-    public String getImagePath() {
-        return imagePath;
-    }
 
-    public void setImagePath(String imagePath) {
-        this.imagePath = imagePath;
-    }
 
     public String getRoutingProfileId() {
         return routingProfileId;
