+++ src/main/java/com/itesm/panoptimize/dto/supervisor/SupervisorCreateDTO.java
@@ -14,7 +14,7 @@ public class SupervisorCreateDTO {
     private String email;
     @NotNull(message = "FullName is required")
     private String fullName;
-    private String imagePath;
+
     @NotNull(message = "RoutingProfileId is required")
     private String routingProfileId;
     private boolean canSwitch = true;
@@ -60,13 +60,7 @@ public class SupervisorCreateDTO {
         this.fullName = fullName;
     }
 
-    public String getImagePath() {
-        return imagePath;
-    }
 
-    public void setImagePath(String imagePath) {
-        this.imagePath = imagePath;
-    }
 
     public String getRoutingProfileId() {
         return routingProfileId;
