+++ src/main/java/com/itesm/panoptimize/dto/supervisor/SupervisorUpdateDTO.java
@@ -18,7 +18,7 @@ public class SupervisorUpdateDTO {
 
     private String fullName;
 
-    private String imagePath;
+
 
     private String routingProfileId;
 
@@ -68,13 +68,7 @@ public class SupervisorUpdateDTO {
         this.fullName = fullName;
     }
 
-    public String getImagePath() {
-        return imagePath;
-    }
 
-    public void setImagePath(String imagePath) {
-        this.imagePath = imagePath;
-    }
 
     public String getRoutingProfileId() {
         return routingProfileId;
