+++ src/main/java/com/itesm/panoptimize/dto/supervisor/SupervisorUserDTO.java
@@ -14,7 +14,7 @@ public class SupervisorUserDTO {
     private String email;
     @NotNull(message = "FullName is required")
     private String fullName;
-    private String imagePath;
+
     @NotNull(message = "RoutingProfileId is required")
     private String routingProfileId;
     private Boolean canSwitch;
@@ -62,13 +62,6 @@ public class SupervisorUserDTO {
         this.fullName = fullName;
     }
 
-    public String getImagePath() {
-        return imagePath;
-    }
-
-    public void setImagePath(String imagePath) {
-        this.imagePath = imagePath;
-    }
 
     public String getRoutingProfileId() {
         return routingProfileId;
