+++ src/main/java/com/itesm/panoptimize/model/User.java
@@ -24,8 +24,6 @@ public class User {
     @Column(name = "full_name", nullable = false)
     private String fullName;
 
-    @Column(name = "image_path")
-    private String imagePath;
 
     @Column(name = "can_switch", nullable = false)
     private Boolean canSwitch = true;
@@ -94,13 +92,7 @@ public class User {
         this.fullName = fullName;
     }
 
-    public String getImagePath() {
-        return imagePath;
-    }
 
-    public void setImagePath(String imagePath) {
-        this.imagePath = imagePath;
-    }
 
     public RoutingProfile getRoutingProfile() {
         return routingProfile;
